requirements.txt
pytorch-lightning==2.0.2
pytorch-pfn-extras
transformers
wandb
black
flake8
isort
mypy
PyYAML
types-PyYAML
click
pulp
ttimer
optuna
japanize_matplotlib